feat: Implement partial rendering of content from Payload Rest API
